ContainerIO: 64-bitized all the setters and getters

+++ Library/include/ContainerIO.h
@@ -179,13 +179,13 @@ extern "C" {
      @abstract                             "Returns the number of audio channels".
      @param                 Audio          "AudioContainer Pointer".
      */
-    uint8_t                 AudioContainer_GetNumChannels(AudioContainer *Audio);
+    uint64_t                AudioContainer_GetNumChannels(AudioContainer *Audio);
     
     /*!
      @abstract                             "Returns the number of audio channels".
      @param                 ChannelMask    "The channel mask".
      */
-    uint8_t                 AudioMask_GetNumChannels(Audio_ChannelMask ChannelMask);
+    uint64_t                AudioMask_GetNumChannels(Audio_ChannelMask ChannelMask);
     
     /*!
      @abstract                             "Returns the number of audio channels".
@@ -352,14 +352,14 @@ extern "C" {
      @remark                               "If there are multiple views, it only returns the number of channels per view".
      @param                 ChannelMask    "The channel mask".
      */
-    uint8_t                 ImageMask_GetNumChannels(Image_ChannelMask ChannelMask);
+    uint64_t                ImageMask_GetNumChannels(Image_ChannelMask ChannelMask);
     
     /*!
      @abstract                             "Returns the number of channels contained in this image".
      @remark                               "If there are multiple views, it only returns the number of channels per view".
      @param                 Image          "ImageContaienr Pointer".
      */
-    uint8_t                 ImageContainer_GetNumChannels(ImageContainer *Image);
+    uint64_t                ImageContainer_GetNumChannels(ImageContainer *Image);
     
     /*!
      @abstract                             "Gets the channel index for the specified mask".
